pyproject.toml
[tool.poetry]
name = "langchain-chatchat"
version = "0.3.1.3"
description = "RAG and Agent applications based on LangChain"
authors = ["AI3721"]
readme = "README.md"

[tool.poetry.scripts]
chatchat = "chatchat.cli:main"

[tool.poetry.dependencies]
python = ">3.9.7,<=3.13"
langchain = "0.3.7"

# api
fastapi = "0.115.5"
uvicorn = "0.32.0"
loguru = "0.7.3"

# client
httpx = "0.27.0"

# web
streamlit = "1.40.1"
streamlit-extras = "0.5.0"
streamlit-chatbox = "1.1.12.post4"
streamlit_antd_components = "0.3.0"

# settings
nltk = "3.9.0"
ruamel_yaml = "0.18.0"
memoization = "0.4.0"
pydantic_settings = "2.5.2"
pydantic = "2.10.2"

# tools
duckduckgo-search = "6.3.6"

# xinference
xinference_client = "1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"